vkgears: do not require vulkan 1.1
We're passing zero as the flags to vkGetDeviceQueue2, meaning we might
as well use vkGetDeviceQueue from Vulkan 1.0 instead. And this seems to
be the only reason we depended on Vulkan 1.1 to start with.

This makes it easier to use vkgears to test early bootstrapped Vulkan
implementations. We still use Vulkan 1.1 if available, because GPU
assisted validation requires it to function.

While we're at it, switch away from VK_MAKE_VERSION(), which is
deprecated. We don't really need it anyway, because we have dedicated
defines for each version.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>
Reviewed-by: Antonino Maniscalco <antomani103@gmail.com>