glsl: remove long-time unsupported demos
These demos depends on ARB_geometry_shader4 support, which was removed
from Mesa back in 2015, with no plans to add it back. This means these
demos can't be run on modern open-source drivers any more, so there's no
point in carrying them around in the tree either.

In addition, porting these to newer GL versions is a lot of work, and
will most likely be more like a complete rewrite. Also, it doesn't seem
like anyone is planning on putting in the work. If someone steps up,
they can dig out the sources from the git history.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>