xeglthreads: Remove #ifdef PTHREADS, fix missing dependency on EGL
Apparently it was stated in c498fc4e40c9b0f62aa14263ebbf79f79a07d949
that PTHREADS is never defined now (that was 13 years ago). Basically
replicated the change. After that it turns out it can't be built since
it actually needs libEGL and that was lost in meson. I guess noone runs
that little program anymore, I decided to give it a bit of love.

Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>