glinfo_common: sort limits a bit
There's a few things that was confusing with the old ordering. First of
all, because GL_EXT_texture_array was in the middle of the GL1.0 section
(which only has extension set to NULL), we ended up making it look like
things like GL_MAX_VIEWPORT_DIMS is a part of GL_EXT_texture_array.
That's not great.

But also, it seems like core GL functionality should come first, as
that's the "least esoteric" information, really.

It's still kinda confusing, because a lot of the extension functionality
has since become core in a later verison, and we're still just printing
these as if they were extension-only. Let's try to deal with that later.

This changes the order of the output, but arguably to be easier to
understand.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>