glinfo_common: add back basic gles-limits
This adds back the basic GLES limits that we dropped earlier. There's a
lot more that we could add here, but I don't really feel like scrathing
that itch right now. This can be tackled later on.

The only thing that's still missing is limits for OpenGL ES 1.x, because
GLAD doesn't support it, and it seems needlessly complicated to support
then. It's also mostly a dead API, so it's unlikely anyone cares.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>