build: be more careful about glvnd
The opengl dependency was added to prefer glvnd when available. But
doing that without specifying the method to pkg-config leads us to
trying to use this on macos, which using the extraframework method will
pick up OpenGL.Framework.

Let's avoid doing that, so macOS users can link to Mesa's libGL again.
While we're at it, turn glvnd into a feature so we can opt in or out on
any platform.

Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>