vkgears: check for sys/time.h
Not all systems have this header, so let's check for it before using it.
We're only using it to provide a timer, which other systems can provide
with other means. So let's add a compile-time error if we have no timer
function.

This is really just preparation for Windows support, where we'll use
other mechanisms for this.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>