wglgears: do not hard-code default window position
Windows has this nice feature where new windows can be given an
automatic position when created, which makes it appear in a bit more
"natural" position than crammed up into the top-left corner of the main
display.

We shouldn't always use the feature, though. If the user fully specified
the geometry, we should for instance let the specified value take
presedence. So let's initialize x to the magical value CW_USEDEFAULT, and
let further processing happen.

It might seem a bit surprising that only x should be set to
CW_USEDEFAULT, but Windows only looks at the x-argument for this behavior.
Setting y to CW_USEDEFAULT has an entirely different meaning, which we
don't want here.

Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>