wglgears: do not change displaymode
There's no good reason we'd want to change the resolution to what it
already was, after all we just queried this from the primary monitor
just before.

The only thing that could have mattered here, would be setting the
display depth to 24 bits. But since we don't need to do this for
windowed mode, there's no good reason we need to do it for fullscreen
either.

I think I was just cargo-culting what I'd been doing for other
full-screen Windows applications, where I actually wanted to change the
resolution etc.

Fixes: 17d95904 ("wglgears: add support for -fullscreen switch")
Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>