wglgears: use up-to-date initial window-size
We skip reacting to the first WM_SIZE that gets issued when the window
is first created, and instead issue our own first reshape call. However,
if a too small window-size has been requested, Windows will end up
giving us a larger window. But since we skipped that first WM_SIZE
event, we're not aware.

So let's query the client-rect of the window explicitly before we issue
the reshape-call.

Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>