vkgears: allow specifying device index
It's been suggested that we should allow users to pick the Vulkan device
to use. So let's add a "-device N" argument, that lets us control the
device index.

Is this the best way of doing this? Dunno, really. We could perhaps
specify the vendorID, deviceID or deviceName from
VkPhysicalDeviceProperties, but we'd either have to specify multiple of
them, or do some fairly hairy guesswork.

So for now, just specifying the index seems like the best option.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>